Mac kext: Scoped timing wrapper and user client for extracting profile data
This adds a scoped time measurement class with splits. Probes are defined in an enum. An external method for extracting the profiling data from the kernel has been added to the logging user client class. The simple log tool has been extended to fetch and dump this data to stdout every 15 seconds. The facility needs to be enabled via a compile time macro and an actual timing sample aggregation library.