Fix for perf tracing integer overflow regression
The `interval` variable is of type `uint64_t`, so the expression `interval * interval` also evaluates to a value of type `uint64_t`, meaning the result of the multiplication will be truncated if `interval` is `1llu << 32` or greater, regardless of the fact that the result is stored in a variable of type `__uint128_t`. Casting one side of the multiplication to the larger type fixes the problem.