Mac: Adds test suite for associative array
Using Xcode's built in XCTest framework, this exercises some of the basic operations of the newly added associative array data structure. Some minor adjustments are made to the header files to make them build in user space, and dummy/test implementations of some of the kext's "library" functions are also provided to make everything work.