Mac kext unit testing: Removed the need for explicit testable symbol list
Using an exports list in the test bundle itself makes dead code stripping work, so the kext-in-userspace can be a static library and doesn't itself need an exports list. This means we don't have to keep adding new functions we want to test to a symbol list.

This change also moves the testability macro to the common header file and adds one for static functions (non-inline), which is tested on yet another function from KauthHandler.cpp.